+++ src/Libraries/SmartStore.Services/Directory/CountryService.cs
@@ -4,78 +4,58 @@ using System.Linq;
 using SmartStore.Core;
 using SmartStore.Core.Caching;
 using SmartStore.Core.Data;
+using SmartStore.Core.Domain.Common;
 using SmartStore.Core.Domain.Directory;
 using SmartStore.Core.Domain.Stores;
 using SmartStore.Core.Events;
+using SmartStore.Templating;
+using SmartStore.Services.Messages;
+using System.Globalization;
 
 namespace SmartStore.Services.Directory
 {
-    /// <summary>
-    /// Country service
-    /// </summary>
     public partial class CountryService : ICountryService
     {
-        #region Constants
         private const string COUNTRIES_ALL_KEY = "SmartStore.country.all-{0}";
         private const string COUNTRIES_BILLING_KEY = "SmartStore.country.billing-{0}";
         private const string COUNTRIES_SHIPPING_KEY = "SmartStore.country.shipping-{0}";
         private const string COUNTRIES_PATTERN_KEY = "SmartStore.country.*";
-        #endregion
-        
-        #region Fields
-        
-        private readonly IRepository<Country> _countryRepository;
-        private readonly IEventPublisher _eventPublisher;
-        private readonly IRequestCache _requestCache;
-		private readonly IStoreContext _storeContext;
+
+		private readonly ICommonServices _services;
+		private readonly IRepository<Country> _countryRepository;
 		private readonly IRepository<StoreMapping> _storeMappingRepository;
+		private readonly ITemplateEngine _templateEngine;
+		private readonly IMessageModelProvider _messageModelProvider;
 
-        #endregion
-
-        #region Ctor
-
-        public CountryService(IRequestCache requestCache,
+		public CountryService(
+			ICommonServices services,
             IRepository<Country> countryRepository,
-            IEventPublisher eventPublisher,
-			IStoreContext storeContext,
-			IRepository<StoreMapping> storeMappingRepository)
+			IRepository<StoreMapping> storeMappingRepository,
+			ITemplateEngine templateEngine,
+			IMessageModelProvider messageModelProvider)
         {
-            _requestCache = requestCache;
             _countryRepository = countryRepository;
-            _eventPublisher = eventPublisher;
-			_storeContext = storeContext;
+			_services = services;
 			_storeMappingRepository = storeMappingRepository;
+			_templateEngine = templateEngine;
+			_messageModelProvider = messageModelProvider;
         }
 
 		public DbQuerySettings QuerySettings { get; set; }
 
-        #endregion
-
-        #region Methods
-
-        /// <summary>
-        /// Deletes a country
-        /// </summary>
-        /// <param name="country">Country</param>
         public virtual void DeleteCountry(Country country)
         {
-            if (country == null)
-                throw new ArgumentNullException("country");
+			Guard.NotNull(country, nameof(country));
 
-            _countryRepository.Delete(country);
-
-            _requestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
+			_countryRepository.Delete(country);
+			
+            _services.RequestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
         }
 
-        /// <summary>
-        /// Gets all countries
-        /// </summary>
-        /// <param name="showHidden">A value indicating whether to show hidden records</param>
-        /// <returns>Country collection</returns>
         public virtual IList<Country> GetAllCountries(bool showHidden = false)
         {
             string key = string.Format(COUNTRIES_ALL_KEY, showHidden);
-            return _requestCache.Get(key, () =>
+            return _services.RequestCache.Get(key, () =>
             {
 				var query = _countryRepository.Table;
 
@@ -86,7 +66,7 @@ namespace SmartStore.Services.Directory
 
 				if (!showHidden && !QuerySettings.IgnoreMultiStore)
 				{
-					var currentStoreId = _storeContext.CurrentStore.Id;
+					var currentStoreId = _services.StoreContext.CurrentStore.Id;
 					query = from c in query
 							join sc in _storeMappingRepository.Table
 							on new { c1 = c.Id, c2 = "Country" } equals new { c1 = sc.EntityId, c2 = sc.EntityName } into c_sm
@@ -107,15 +87,10 @@ namespace SmartStore.Services.Directory
             });
         }
 
-        /// <summary>
-        /// Gets all countries that allow billing
-        /// </summary>
-        /// <param name="showHidden">A value indicating whether to show hidden records</param>
-        /// <returns>Country collection</returns>
         public virtual IList<Country> GetAllCountriesForBilling(bool showHidden = false)
         {
             string key = string.Format(COUNTRIES_BILLING_KEY, showHidden);
-            return _requestCache.Get(key, () =>
+            return _services.RequestCache.Get(key, () =>
             {
 				var allCountries = GetAllCountries(showHidden);
 
@@ -124,15 +99,10 @@ namespace SmartStore.Services.Directory
             });
         }
 
-        /// <summary>
-        /// Gets all countries that allow shipping
-        /// </summary>
-        /// <param name="showHidden">A value indicating whether to show hidden records</param>
-        /// <returns>Country collection</returns>
         public virtual IList<Country> GetAllCountriesForShipping(bool showHidden = false)
         {
             string key = string.Format(COUNTRIES_SHIPPING_KEY, showHidden);
-            return _requestCache.Get(key, () =>
+            return _services.RequestCache.Get(key, () =>
             {			
 				var allCountries = GetAllCountries(showHidden);
 
@@ -141,11 +111,6 @@ namespace SmartStore.Services.Directory
             });
         }
 
-        /// <summary>
-        /// Gets a country 
-        /// </summary>
-        /// <param name="countryId">Country identifier</param>
-        /// <returns>Country</returns>
         public virtual Country GetCountryById(int countryId)
         {
             if (countryId == 0)
@@ -154,11 +119,6 @@ namespace SmartStore.Services.Directory
             return _countryRepository.GetById(countryId);
         }
 
-		/// <summary>
-		/// Gets a country by two or three letter ISO code
-		/// </summary>
-		/// <param name="letterIsoCode">Country two or three letter ISO code</param>
-		/// <returns>Country</returns>
 		public virtual Country GetCountryByTwoOrThreeLetterIsoCode(string letterIsoCode)
 		{
 			if (letterIsoCode.HasValue())
@@ -171,11 +131,6 @@ namespace SmartStore.Services.Directory
 			return null;
 		}
 
-        /// <summary>
-        /// Gets a country by two letter ISO code
-        /// </summary>
-        /// <param name="twoLetterIsoCode">Country two letter ISO code</param>
-        /// <returns>Country</returns>
         public virtual Country GetCountryByTwoLetterIsoCode(string twoLetterIsoCode)
         {
 			if (twoLetterIsoCode.IsEmpty())
@@ -189,11 +144,6 @@ namespace SmartStore.Services.Directory
             return country;
         }
 
-        /// <summary>
-        /// Gets a country by three letter ISO code
-        /// </summary>
-        /// <param name="threeLetterIsoCode">Country three letter ISO code</param>
-        /// <returns>Country</returns>
         public virtual Country GetCountryByThreeLetterIsoCode(string threeLetterIsoCode)
         {
 			if (threeLetterIsoCode.IsEmpty())
@@ -207,34 +157,57 @@ namespace SmartStore.Services.Directory
             return country;
         }
 
-        /// <summary>
-        /// Inserts a country
-        /// </summary>
-        /// <param name="country">Country</param>
         public virtual void InsertCountry(Country country)
         {
-            if (country == null)
-                throw new ArgumentNullException("country");
+			Guard.NotNull(country, nameof(country));
 
-            _countryRepository.Insert(country);
+			_countryRepository.Insert(country);
 
-            _requestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
+			_services.RequestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
         }
 
-        /// <summary>
-        /// Updates the country
-        /// </summary>
-        /// <param name="country">Country</param>
         public virtual void UpdateCountry(Country country)
         {
-            if (country == null)
-                throw new ArgumentNullException("country");
+			Guard.NotNull(country, nameof(country));
 
             _countryRepository.Update(country);
 
-            _requestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
+			_services.RequestCache.RemoveByPattern(COUNTRIES_PATTERN_KEY);
         }
 
-        #endregion
-    }
+		public virtual string FormatAddress(Address address, bool newLineToBr = false)
+		{
+			Guard.NotNull(address, nameof(address));
+
+			var template = address.Country?.AddressFormat.NullEmpty() ?? Address.DefaultAddressFormat;
+
+			var messageContext = new MessageContext
+			{
+				Language = _services.WorkContext.WorkingLanguage,
+				Store = _services.StoreContext.CurrentStore,
+				Model = new TemplateModel()
+			};
+
+			_messageModelProvider.AddModelPart(address, messageContext, "Address");
+			var model = messageContext.Model["Address"];
+			var result = _templateEngine.Render(template, model, messageContext.FormatProvider).Compact(true);
+
+			if (newLineToBr)
+			{
+				result = Core.Html.HtmlUtils.ConvertPlainTextToHtml(result);
+			}		
+
+			return result;
+		}
+
+		public virtual string FormatAddress(object address, Country country = null, IFormatProvider formatProvider = null)
+		{
+			Guard.NotNull(address, nameof(address));
+
+			var template = country?.AddressFormat.NullEmpty() ?? Address.DefaultAddressFormat;
+			var result = _templateEngine.Render(template, address, formatProvider ?? CultureInfo.CurrentCulture).Compact(true);
+
+			return result;
+		}
+	}
 }