+++ src/Libraries/SmartStore.Services/Messages/MessageContext.cs
@@ -1,15 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Localization;
 using SmartStore.Core.Domain.Messages;
 using SmartStore.Core.Domain.Stores;
 using SmartStore.Core.Email;
+using SmartStore.Core.Localization;
 
 namespace SmartStore.Services.Messages
 {
 	public class MessageContext
 	{
+		private IFormatProvider _formatProvider;
+
 		public MessageTemplate MessageTemplate { get; set; }
 		public string MessageTemplateName { get; set; }
 
@@ -31,7 +35,39 @@ namespace SmartStore.Services.Messages
 		public Uri BaseUri { get; set; }
 
 		public TemplateModel Model { get; set; }
-		public IFormatProvider FormatProvider { get; set; }
+
+		public IFormatProvider FormatProvider
+		{
+			get
+			{
+				if (_formatProvider == null)
+				{
+					var culture = this.Language?.LanguageCulture;
+					if (culture != null && LocalizationHelper.IsValidCultureCode(culture))
+					{
+						_formatProvider = CultureInfo.GetCultureInfo(culture);
+					}
+				}
+
+				return _formatProvider ?? CultureInfo.CurrentCulture;
+			}
+			set
+			{
+				_formatProvider = value;
+			}
+		}
+
+		private IFormatProvider GetFormatProvider(MessageContext messageContext)
+		{
+			var culture = messageContext.Language.LanguageCulture;
+
+			if (LocalizationHelper.IsValidCultureCode(culture))
+			{
+				return CultureInfo.GetCultureInfo(culture);
+			}
+
+			return CultureInfo.CurrentCulture;
+		}
 
 		public static MessageContext Create(string messageTemplateName, int languageId, int? storeId = null, Customer customer = null)
 		{