+++ src/Presentation/SmartStore.Web.Framework/Extensions/HttpExtensions.cs
@@ -192,9 +192,11 @@ namespace SmartStore
 
 			if (value.HasValue() && cookie == null)
 			{
-				cookie = new HttpCookie("sm.UserThemeChoice");
-				cookie.HttpOnly = true;
-				cookie.Expires = DateTime.UtcNow.AddYears(1);					
+				cookie = new HttpCookie("sm.UserThemeChoice")
+				{
+					HttpOnly = true,
+					Expires = DateTime.UtcNow.AddYears(1)
+				};				
 			}
 
 			if (value.HasValue())
@@ -223,8 +225,7 @@ namespace SmartStore
 			
 			if (cookie == null && createIfMissing)
 			{
-				cookie = new HttpCookie("sm.PreviewModeOverrides");
-				cookie.HttpOnly = true;
+				cookie = new HttpCookie("sm.PreviewModeOverrides") { HttpOnly = true };
 				context.Request.Cookies.Set(cookie);
 			}