+++ src/Presentation/SmartStore.Web.Framework/Templating/Liquid/Drops/ObjectDrop.cs
@@ -47,5 +47,10 @@ namespace SmartStore.Templating.Liquid
 
 			return null;
 		}
+
+		public override object GetWrappedObject()
+		{
+			return _data;
+		}
 	}
 }