+++ src/Presentation/SmartStore.Web.Framework/Templating/Liquid/Drops/TestDrop.cs
@@ -30,6 +30,11 @@ namespace SmartStore.Templating.Liquid
 			get => _type.Name;
 		}
 
+		public object GetWrappedObject()
+		{
+			return _entity;
+		}
+
 		public bool ContainsKey(object key)
 		{
 			return true;