+++ src/Presentation/SmartStore.Web/Administration/Models/Directory/CountryModel.cs
@@ -50,7 +50,10 @@ namespace SmartStore.Admin.Models.Directory
         [SmartResourceDisplayName("Admin.Configuration.Countries.Fields.NumberOfStates")]
         public int NumberOfStates { get; set; }
 
-        public IList<CountryLocalizedModel> Locales { get; set; }
+		[SmartResourceDisplayName("Admin.Configuration.Countries.Fields.AddressFormat")]
+		public string AddressFormat { get; set; }
+
+		public IList<CountryLocalizedModel> Locales { get; set; }
 
 		[SmartResourceDisplayName("Admin.Common.Store.LimitedTo")]
 		public bool LimitedToStores { get; set; }