+++ src/Presentation/SmartStore.Web/Extensions/MappingExtensions.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
+using SmartStore.ComponentModel;
 using SmartStore.Core.Domain.Catalog;
 using SmartStore.Core.Domain.Common;
 using SmartStore.Core.Domain.Directory;
@@ -67,7 +68,8 @@ namespace SmartStore.Web
         /// <param name="stateProvinceService">State service (used to prepare a select list). null to don't prepare the list.</param>
         /// <param name="loadCountries">A function to load countries  (used to prepare a select list). null to don't prepare the list.</param>
         public static void PrepareModel(this AddressModel model,
-            Address address, bool excludeProperties,
+            Address address, 
+			bool excludeProperties,
             AddressSettings addressSettings,
             ILocalizationService localizationService = null,
             IStateProvinceService stateProvinceService = null,
@@ -78,34 +80,18 @@ namespace SmartStore.Web
 
             if (addressSettings == null)
                 throw new ArgumentNullException("addressSettings");
-
+			
             if (!excludeProperties && address != null)
             {
-                model.Id = address.Id;
-                model.Salutation = address.Salutation;
-                model.Title = address.Title;
-                model.FirstName = address.FirstName;
-                model.LastName = address.LastName;
-                model.Email = address.Email;
-                model.EmailMatch = address.Email;
-                model.Company = address.Company;
-                model.CountryId = address.CountryId;
-                model.CountryName = address.Country != null
-                    ? address.Country.GetLocalized(x => x.Name)
-                    : null;
-                model.StateProvinceId = address.StateProvinceId;
-                model.StateProvinceName = address.StateProvince != null
-                    ? address.StateProvince.GetLocalized(x => x.Name)
-                    : null;
-                model.City = address.City;
-                model.Address1 = address.Address1;
-                model.Address2 = address.Address2;
-                model.ZipPostalCode = address.ZipPostalCode;
-                model.PhoneNumber = address.PhoneNumber;
-                model.FaxNumber = address.FaxNumber;
-            }
+				MiniMapper.Map(address, model);
 
-            //countries and states
+				model.EmailMatch = address.Email;
+				model.CountryName = address.Country?.GetLocalized(x => x.Name);
+				model.StateProvinceName = address.StateProvince.GetLocalized(x => x.Name);
+				model.FormattedAddress = Core.Infrastructure.EngineContext.Current.Resolve<ICountryService>().FormatAddress(address, true);
+			}
+
+            // countries and states
             if (addressSettings.CountryEnabled && loadCountries != null)
             {
                 if (localizationService == null)
@@ -114,7 +100,7 @@ namespace SmartStore.Web
                 model.AvailableCountries.Add(new SelectListItem() { Text = localizationService.GetResource("Address.SelectCountry"), Value = "0" });
                 foreach (var c in loadCountries())
                 {
-                    model.AvailableCountries.Add(new SelectListItem()
+                    model.AvailableCountries.Add(new SelectListItem
                     {
                         Text = c.GetLocalized(x => x.Name),
                         Value = c.Id.ToString(),
@@ -124,12 +110,12 @@ namespace SmartStore.Web
 
                 if (addressSettings.StateProvinceEnabled)
                 {
-                    //states
+                    // states
                     if (stateProvinceService == null)
                         throw new ArgumentNullException("stateProvinceService");
 
                     var states = stateProvinceService
-                        .GetStateProvincesByCountryId(model.CountryId.HasValue ? model.CountryId.Value : 0)
+                        .GetStateProvincesByCountryId(model.CountryId ?? 0)
                         .ToList();
                     if (states.Count > 0)
                     {
@@ -164,26 +150,8 @@ namespace SmartStore.Web
                 }
             }
 
-            //form fields
-            model.ValidateEmailAddress = addressSettings.ValidateEmailAddress;
-            model.SalutationEnabled = addressSettings.SalutationEnabled;
-            model.TitleEnabled = addressSettings.TitleEnabled;
-            model.CompanyEnabled = addressSettings.CompanyEnabled;
-            model.CompanyRequired = addressSettings.CompanyRequired;
-            model.StreetAddressEnabled = addressSettings.StreetAddressEnabled;
-            model.StreetAddressRequired = addressSettings.StreetAddressRequired;
-            model.StreetAddress2Enabled = addressSettings.StreetAddress2Enabled;
-            model.StreetAddress2Required = addressSettings.StreetAddress2Required;
-            model.ZipPostalCodeEnabled = addressSettings.ZipPostalCodeEnabled;
-            model.ZipPostalCodeRequired = addressSettings.ZipPostalCodeRequired;
-            model.CityEnabled = addressSettings.CityEnabled;
-            model.CityRequired = addressSettings.CityRequired;
-            model.CountryEnabled = addressSettings.CountryEnabled;
-            model.StateProvinceEnabled = addressSettings.StateProvinceEnabled;
-            model.PhoneEnabled = addressSettings.PhoneEnabled;
-            model.PhoneRequired = addressSettings.PhoneRequired;
-            model.FaxEnabled = addressSettings.FaxEnabled;
-            model.FaxRequired = addressSettings.FaxRequired;
+			// Form fields
+			MiniMapper.Map(addressSettings, model);
         }
 
         public static Address ToEntity(this AddressModel model)